First implementation of 8.4.2.
	modified:   Configuration/DBScript/Phase8Sprint4.sql
	modified:   idas.chox.data/src/main/java/idas/chox/data/events/EventGenerator.java
	modified:   idas.chox.data/src/main/java/idas/chox/data/services/BaseDataService.java
	modified:   idas.chox.data/src/main/java/idas/chox/data/services/ClaimServiceImpl.java
	modified:   idas.chox.data/src/test/resources/import.sql
	modified:   idas.chox.service/src/main/java/idas/chox/service/security/ApplicationAccessibility.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ActivityEvent.java
	modified:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/ActivityEventGenerator.java
	new file:   idas.chox.service/src/main/java/idas/chox/service/workflow/activities/SwitchCho.java
	new file:   idas.chox.service/src/test/java/idas/chox/service/workflow/activities/SwitchChoTest.java
	modified:   idas.chox.service/src/test/resources/applicationContext-Workflow-test.xml
	modified:   idas.chox.service/src/test/resources/import.sql
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/BaseAction.java
	modified:   idas.chox.web/src/main/java/idas/chox/web/actions/ClaimAction.java
	modified:   idas.chox.web/src/main/resources/applicationContext-Workflow.xml
	modified:   idas.chox.web/src/main/resources/struts.xml
	modified:   idas.chox.web/src/main/webapp/jsp/claim_detail.jsp

+++ idas.chox.data/src/main/java/idas/chox/data/events/EventGenerator.java
@@ -1,9 +1,10 @@
 package idas.chox.data.events;
 
-import idas.chox.core.model.Claim;
-import idas.chox.events.EventRegister;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import idas.chox.core.model.Claim;
+import idas.chox.events.EventRegister;
 
 /**
  *
@@ -21,15 +22,15 @@ public class EventGenerator {
     public void startEvent(Claim claim, String name, int id, boolean insurerOnly, boolean choOnly) throws Exception {
         int claimId = -1;
         if (claim.getId() != null) {
-            claimId = claim.getId().intValue();
+            claimId = claim.getId();
         }
         int insurerId = 0;
         if (!choOnly) {
-            insurerId = claim.getInsurer().getId().intValue();
+            insurerId = claim.getInsurer().getId();
         }
         int choId = 0;
         if (!insurerOnly) {
-            choId = claim.getChorganisation().getId().intValue();
+            choId = claim.getChorganisation().getId();
         }
         eventRegister.startEvent(name, id, insurerId, choId, claimId, claim.getClaimType().ordinal());
         eventRegister.addParameter("insurerName", claim.getInsurer().getName());
